Crude fix for 'dirty' bash prompt flag.
Add " * " to the bash prompt if you are in a dirty git working
directory or if your current branch is ahead of or behind the remote
branch it is tracking.

The method for detecting a clean working directory is crude:
a hard-coded match to an expected text of "nothing to commit,
working directory clean" in the first lines of output from
'git branch'.  But the previous version didn't work at all.