Add "dirty" marker to git branch in prompt.
Uses the output of `git status --short`, which seems to include a code
of one letter for each file in the index or "??" for an untracked
file. This commit concatenates those letters and "??"s as the marker
in the prompt for a "dirty" git status. The marker shows next to the
git branch name when in a git project directory.