Install a modified version of nzig/fzf.fish plugin
nzig/fzf.fish has enhancements to PatrickF1's version, including
an fzf mode to search for git branches. To nzig's code, I added a
default keybinding for the search git branches feature (Alt+Ctrl+B),
enabled configuration of the keybindings for that mode as well as
for magic mode, and updated the help information to cover both
magic mode and git branches mode. See pull request
* https://github.com/nzig/fzf.fish/pull/1

This commit adds the hand-modified version of the plugin to dotfiles.