Add vim_in_color function. See escape seq as color
Open a file in a read-only pane (would prefer the whole window) with
all color escape sequences rendering text in the color specified.
Great for viewing log or other files with embedded color escape
sequences, unless you prefer
    ^[[36mcarp_1 |^[[0m 03:30:47 ^[[36mI^[[0m [9:5100] ^[[36mcarp^[[0m

It is slow, and misleading because it opens the main vim window right
away but opens a pane inside the window later, whenever the file
rendering is ready (it seems). For example, it takes about 10 seconds
to open the color pane on a 57,000,000-character file, during which
time you have no indication that something is happening.